visual_tagger_ai/backend/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
Pillow 